+++ website/src/routes/list/+page.svelte
@@ -67,9 +67,9 @@
 
   <VariantExample
     code={`
-<ul ...> 
+<Box ...> 
   <slot />
-</ul>
+</Box>
 `}
   />